+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/MrpServiceImpl.java
@@ -868,23 +868,23 @@ public class MrpServiceImpl implements MrpService {
 
   protected List<StockLocation> getAllLocationAndSubLocation(StockLocation stockLocation) {
 
-    List<StockLocation> subLocationList =
+    List<StockLocation> resultList = new ArrayList<>();
+
+    for (StockLocation subLocation :
         stockLocationRepository
             .all()
             .filter(
                 "self.parentStockLocation.id = :stockLocationId AND self.typeSelect != :virtual")
             .bind("stockLocationId", stockLocation.getId())
             .bind("virtual", StockLocationRepository.TYPE_VIRTUAL)
-            .fetch();
+            .fetch()) {
 
-    for (StockLocation subLocation : subLocationList) {
-
-      subLocationList.addAll(this.getAllLocationAndSubLocation(subLocation));
+      resultList.addAll(this.getAllLocationAndSubLocation(subLocation));
     }
 
-    subLocationList.add(stockLocation);
+    resultList.add(stockLocation);
 
-    return subLocationList;
+    return resultList;
   }
 
   @Override