+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/PurchaseOrderServiceSupplychainImpl.java
@@ -227,11 +227,15 @@ public class PurchaseOrderServiceSupplychainImpl extends PurchaseOrderServiceImp
                   null,
                   StockMoveRepository.TYPE_INCOMING);
 
-      stockMove.setPurchaseOrder(purchaseOrder);
+      stockMove.setOriginId(purchaseOrder.getId());
+      stockMove.setOriginTypeSelect(StockMoveRepository.ORIGIN_PURCHASE_ORDER);
+      stockMove.setOrigin(purchaseOrder.getPurchaseOrderSeq());
       stockMove.setEstimatedDate(purchaseOrder.getDeliveryDate());
       stockMove.setTradingName(purchaseOrder.getTradingName());
 
-      qualityStockMove.setPurchaseOrder(purchaseOrder);
+      qualityStockMove.setOriginId(purchaseOrder.getId());
+      qualityStockMove.setOriginTypeSelect(StockMoveRepository.ORIGIN_PURCHASE_ORDER);
+      qualityStockMove.setOrigin(purchaseOrder.getPurchaseOrderSeq());
       qualityStockMove.setEstimatedDate(purchaseOrder.getDeliveryDate());
       qualityStockMove.setTradingName(purchaseOrder.getTradingName());
 
@@ -359,7 +363,8 @@ public class PurchaseOrderServiceSupplychainImpl extends PurchaseOrderServiceImp
         Beans.get(StockMoveRepository.class)
             .all()
             .filter(
-                "self.purchaseOrder.id = ? AND self.statusSelect <> ?",
+                "self.originTypeSelect LIKE ? AND self.originId = ? AND self.statusSelect <> ?",
+                StockMoveRepository.ORIGIN_PURCHASE_ORDER,
                 purchaseOrderId,
                 StockMoveRepository.STATUS_CANCELED)
             .count();