+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockMoveServiceSupplychainImpl.java
@@ -56,6 +56,7 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
   @Inject private AppSupplychainService appSupplyChainService;
 
   @Inject protected PurchaseOrderRepository purchaseOrderRepo;
+  @Inject protected SaleOrderRepository saleOrderRepo;
 
   @Inject private PurchaseOrderServiceSupplychainImpl purchaseOrderServiceSupplychain;
 
@@ -83,10 +84,11 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
         "Réalisation du mouvement de stock : {} ", new Object[] {stockMove.getStockMoveSeq()});
     String newStockSeq = super.realize(stockMove, check);
     AppSupplychain appSupplychain = appSupplyChainService.getAppSupplychain();
-    if (stockMove.getSaleOrder() != null) {
+
+    if (StockMoveRepository.ORIGIN_SALE_ORDER.equals(stockMove.getOriginTypeSelect())) {
       updateSaleOrderLines(stockMove, true);
       // Update linked saleOrder delivery state depending on BackOrder's existence
-      SaleOrder saleOrder = stockMove.getSaleOrder();
+      SaleOrder saleOrder = saleOrderRepo.find(stockMove.getOriginId());
       if (newStockSeq != null) {
         saleOrder.setDeliveryState(SaleOrderRepository.DELIVERY_STATE_PARTIALLY_DELIVERED);
       } else {
@@ -99,9 +101,9 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
       }
 
       Beans.get(SaleOrderRepository.class).save(saleOrder);
-    } else if (stockMove.getPurchaseOrder() != null) {
+    } else if (StockMoveRepository.ORIGIN_PURCHASE_ORDER.equals(stockMove.getOriginTypeSelect())) {
       // Update linked purchaseOrder receipt state depending on BackOrder's existence
-      PurchaseOrder purchaseOrder = stockMove.getPurchaseOrder();
+      PurchaseOrder purchaseOrder = purchaseOrderRepo.find(stockMove.getOriginId());
       if (newStockSeq != null) {
         purchaseOrder.setReceiptState(IPurchaseOrder.STATE_PARTIALLY_RECEIVED);
       } else {
@@ -121,11 +123,11 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
   @Transactional(rollbackOn = {AxelorException.class, Exception.class})
   public void cancel(StockMove stockMove) throws AxelorException {
     if (stockMove.getStatusSelect() == StockMoveRepository.STATUS_REALIZED) {
-      if (stockMove.getSaleOrder() != null) {
+      if (StockMoveRepository.ORIGIN_SALE_ORDER.equals(stockMove.getOriginTypeSelect())) {
         updateSaleOrderOnCancel(stockMove);
       }
-      if (stockMove.getPurchaseOrder() != null) {
-        PurchaseOrder purchaseOrder = purchaseOrderRepo.find(stockMove.getPurchaseOrder().getId());
+      if (StockMoveRepository.ORIGIN_PURCHASE_ORDER.equals(stockMove.getOriginTypeSelect())) {
+        PurchaseOrder purchaseOrder = purchaseOrderRepo.find(stockMove.getOriginId());
         purchaseOrderServiceSupplychain.updatePurchaseOrderOnCancel(stockMove, purchaseOrder);
       }
     }
@@ -134,9 +136,16 @@ public class StockMoveServiceSupplychainImpl extends StockMoveServiceImpl
 
   @Transactional(rollbackOn = {Exception.class})
   public void updateSaleOrderOnCancel(StockMove stockMove) {
-    SaleOrder so = Beans.get(SaleOrderRepository.class).find(stockMove.getSaleOrder().getId());
+    SaleOrder so = saleOrderRepo.find(stockMove.getOriginId());
 
-    List<StockMove> stockMoveList = stockMoveRepo.all().filter("self.saleOrder = ?1", so).fetch();
+    List<StockMove> stockMoveList =
+        stockMoveRepo
+            .all()
+            .filter(
+                "self.originId = ?1 AND self.originTypeSelect = ?2",
+                so.getId(),
+                StockMoveRepository.ORIGIN_SALE_ORDER)
+            .fetch();
     so.setDeliveryState(SaleOrderRepository.DELIVERY_STATE_NOT_DELIVERED);
     for (StockMove stock : stockMoveList) {
       if (stock.getStatusSelect() != StockMoveRepository.STATUS_CANCELED