+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - Set project on purchase order and purchase order lines when generating it from a sale order.
 - Improve account management views.
 - Rework ManufOrder printing.
+- Add new Origin field in stock moves and remove fields sale order and purchase order.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.
@@ -85,6 +86,7 @@
 - Added new filter for invoice project (for SaleOrderLine, PurchaseOrderLine, and ExpenseLine o2m)
 - Project: Fix financial report view.
 - Fix stock move generation with pack lines from sale order
+- MRP: Fix concurrent modification exception.
 
 ## [Unreleased 5.0.5]	
 ##Features